requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0

# Browser automation: handles Twitch OAuth login flow + API discovery
playwright>=1.43.0
